README.md
# choose_your_own_adventure_game
Saving Princess Emily

Introduction
'Saving Princess Emily' (SPE) has been designed as a choose your own adventure game with a medieval theme, where players are able to select the direction they want the story to go in with a personalised character name. Throughout the game, players select their character, directions, decisions and attacking strategies as they encounter scenarios with royalty, creatures and crossroads. 

Purpose:
The game was developed as part of an introduction to programming course the developer undertook in 2020, to demonstrate technical sophistication and an understanding of terminal processes.

Language:
Ruby

How to run the project:
This program does not require installation.

Target audience:
SPE was designed for teenagers and children (10 - 18 years old), although all ages are encouraged if they wish to delve into the medieval world and explore their imagination. The target audience was selected as the story line is child-friendly, contains age appropriate vocabulary and is thought that it would be enjoyed the most by this group.

Github:
https://github.com/CaladenTemmo/choose_your_own_adventure_game

Features 

The pathway system: which will allow players to choose between different encounters.
Personalised characters: that enables players to decide between 2 different characters and give them a name.
Health and damage system: provides an objective for the players to finish the game and avoid losing (death).
The user will find out how to use each feature when the game explains and tells the player what to press or do for each feature.
The player will interact with each feature by entering the player's name or choosing a path and even features such as health and damage.


Status Updates

I changed the health system from its own method and added it into the monstera method either sea monster or trolls method.
I had an issue with the damage system adding up properly instead of saying the amount of damage you would do would just keep adding up the amount of damage you do.
I had an issue with the pathway system; it wouldn't go from one file to the next.
I've just started on the character choice but am having a problem with the characters
I'm halfway through adding the damage system but it's not adding up correctly.
I've just tested my character selection as well as a pathway to choose and both are working.
I've just tested my health and damage system and both are working with no errors.






Help file
To install go to github (https://github.com/CaladenTemmo/choose_your_own_adventure_game) then press the green button that says clone or download
You will need tty-prompy and artii gems
There are features such as choosing your character and naming your character, choosing a pathway to decide which battle you will go into and a health and damage fighting system.




Gems 

require "tty-prompt"
require "artii"
require "colorize"



Implementation
Trello:
https://trello.com/b/RPU7rAxC/terminal-app




Day 1

I start writing up a story for my choose your own adventure game 
I start working on basic code for the start of the game and developing the start of the game
Looking into any gems i might need

Day 2

I finish working on the intro to the game as well as starting the code for being able to choose different characters
I add in options for you to choose to accept or deny the king in the game
Coding in the knights story along with very early start to the story for the wizard

Day 3

I finish up the code for the knight and the basic story for the wizard which is still in its early days
Now i add a location in the docks and the bridge

Day 4

I add in the sea monster and start on some code for it as well as the troll
I add a health and damage system
I add a basic fighting system to the game where you can kill the sea monster or die

Day 5

I add you being able to save the princess if you defeat the sea monster
I better my fighting system so now you are able to have multiple attacks that do different amount of damage
I add in a random system to determine how much damage you do deciding between number e.g. 15 to 25 damage

Day 6

I add the ability to name you character
And I add the wizard and trolls to the game

Day 7

I clean up all the text
I run through the game to see if there are anymore errors
And check for any problems